Update patch set 2
Patch Set 2:

I believe Jess was correct.  The released version in the R1 is 1.1.0.  So we should bump to 1.2.0.  (it is what's in the version.properties counts what doing the release build).  We should up the versions in pom's to 1.2.0 too.

Patch-set: 2
Reviewer: Gerrit User 51 <51@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0