Update patch set 3
Patch Set 3:

(7 comments)

@jdudycz In reply to general comment - yes, this commit is supposed to include changes made in whole "HV-VES: Move ENVs to CBS" task, and as such it should be merged after finishing all sub-tasks.

However as we were little blocked with tasks I decided to start this tracking-changeset and also include there information that we are aiming for.

I tried to not be very precise in mechanisms that are not yet implemented, however I find information about them helpful in overall "view of docs".

Patch-set: 3