Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (2 comments)
> 
> minor updates needed

Vijay, 

Thanks for your review.
I suggest the following phrasing, starting with line 46 (as it is now) and then going on with "Validate blueprint and the next lines as they are now:

BBS-ep can be dynamically deployed in DCAE’s Cloudify environment via its blueprint deployment artifact.

- Blueprint can be found in

    https://gerrit.onap.org/r/gitweb?p=dcaegen2/platform/blueprints.git;a=blob_plain;f=blueprints/k8s-bbs-event-processor.yaml;hb=refs/heads/master

- Blueprint can also be retrieved from

    https://git.onap.org/dcaegen2/platform/blueprints/tree/blueprints/k8s-bbs-event-processor.yaml

Steps to deploy are shown below

- Enter the Bootstrap POD
- Validate blueprint
...

Patch-set: 2