Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> It's my understanding that the /v1 paths should not be removed
> 
> and that all new paths should be available under both the /v2 level in addition to their being available at the top level.
> 
> Is there a reason that these updates were not done that way?

I am looking for some confirmation on this. The only api call being continued in the new version is the healthcheck and that is unchanged. Therefore, I don't believe that this needs versioning.
All other api calls are new calls being implemented for the first time in this release.
As I say I'm trying to find some clear documentation on this case. I'm open to any suggestions.

Patch-set: 2