Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (1 comment)
> 
> It's my understanding that the /v1 paths should not be removed
> 
> and that all new paths should be available under both the /v2 level in addition to their being available at the top level.
> 
> Is there a reason that these updates were not done that way?

After talking to the team it would appear that no API has ever been published in the PMSH documentation before. As such we don't believe that we need to version here as this is the first official API that will be published and documented for the PMSH product. We can prefix v1 in the URI if necessary but don't believe any other action should be needed. Is this an acceptable approach? We are hoping to have this resolved and merged for M2.

Patch-set: 2