Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1: Code-Review+1
> 
> The validation script is great but why must someone fix other peoples errors?
> I made a change to the deployment.rst file but did not add or remove the newline form the file ending.
> Bit frustrating if I am honest.
> I am all for "clean as you go" when coding but maybe not in this instance.

The check for "is there a newline at the end of the file" is run on all modified source files, which includes .rst files. So this is a valid complaint for deployment.rst. Just add a newline to the end of the file.

Patch-set: 1
Reviewer: Gerrit User 105 <105@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1